\name{met-package}
\alias{met-package}
\alias{met}
\docType{package}
\title{
\packageTitle{met}
}
\description{
\packageDescription{met}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{met}
\packageIndices{met}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{met}

Maintainer: \packageMaintainer{met}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
